expand index to 128
This requires reworking the encoding to get an extra bit in the reference chunk.
I decided to drop the small and medium modification chunk. It also got an extra
bit in the small run chunk.

Quick experiments suggest it's not as big of a win (if it's a win at all) as I was
hoping for. Archiving it here, to return to after closer examination of what chunk
types are used most.