don't use the 3-byte mod chunk, when a 2-byte sub chunk would suffice
If the difference in 3/4 components is zero, it doesn't matter if the fourth diff
would fit in the "large" mod chunk. It's more efficient to encode the pixel value
as a 2-byte substitution of a single component.

This change has no effect on the mario world map sample, but saves 160 bytes on the
phone screenshot example.

Also included is a quick tool I wrote to find the first differing chunk(s) between
two streams, which helped to debug this change.