remove wrap_sum
The math that wrap_sum was doing is the same as just taking the lowest eight bits
of the computed value. Putting the value in a pixel binary is going to take the
lowest eight bits. The new code spends half as much time in mod_pixel, according to
fprof.