merge encode_image and encode_pixel into encode_loop
The only reason for encode_image to exist is that Channels, one of its arguments,
can't be used in a match on Image, one of its other arguments, in the function
heading, because it isn't yet bound.

This change moves all of the code from encode_pixel into an encode_loop function
body, defined by a macro. The macro's argument is Channels. So, when the preprocess
produces the code, instead of Channels in the function head, there is a litteral 3
or 4.

Speed improvements of this change are inconclusive enough at this point that I
don't want to include something that is so convoluted on main just yet. Although it
would be a lot of code duplication, writing separate encode_loop3 and encode_loop4
functions would probably be clearer. That would also allow other 3-channel
optimizations, like dropping absolutely everything that checks on alpha
differences.